refactor(stft): update window implementation to use f64 precision instead of f32

+++ crates/bunsen/src/ops/signal/stft_window.rs
@@ -53,7 +53,7 @@ impl SamplingWindowBuilder for StftWindowConfig {
     fn to_vec_window(
         &self,
         size: usize,
-    ) -> Vec<f32> {
+    ) -> Vec<f64> {
         match self {
             Self::Ones => vec![1.0; size],
             Self::CosineWindow(cfg) => cfg.to_vec_window(size),
